Simplify the handling of slices
Revert to previous approach to slice handling while retaining the split
tag approach.

Checking for split tags now properly queries the template for inclusion
of the split tag or the args function. We still have a problem where
users might have a pair of adjacent pipes in a filename, this should be
rare.