Use null marks to split arguments
Well that was an interesting little diversion.

Incorporate a split function which inserts a null byte into a pipeline
and split on that mark. We can reasonably safely assume that null bytes
are not going to be present in a file name for most file systems. For
this reason we can also remove the template querying.